+++ frontend/src/app/modules/page/other-parts/other-parts.module.ts
@@ -30,9 +30,11 @@ import { OtherPartsService } from './core/other-parts.service';
 import { OtherPartsState } from './core/other-parts.state';
 import { OtherPartsRoutingModule } from './other-parts.routing';
 import { OtherPartsComponent } from './presentation/other-parts.component';
+import { SupplierPartsComponent } from '@page/other-parts/presentation/supplier-parts/supplier-parts.component';
+import { CustomerPartsComponent } from './presentation/customer-parts/customer-parts.component';
 
 @NgModule({
-  declarations: [OtherPartsComponent],
+  declarations: [OtherPartsComponent, SupplierPartsComponent, CustomerPartsComponent],
   imports: [CommonModule, TemplateModule, SharedModule, OtherPartsRoutingModule, PartDetailsModule],
   providers: [
     OtherPartsState,