+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased - x.x.x]
 
+### Added
+
+- Added option to hide "Investigation for components" in part-detail.component.html
+
+### Changed
+
+- Refactored other-parts.component.html - split into new components: supplier-parts.component.html and customer-parts.component.html
+
+### Removed
+- Removed selection column on Customer Parts page
+
 ## [3.4.0] - 2023-05-11
 
 ### Added
@@ -30,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Upgraded karma dependencies (use engine.io@^6.4.2 to solve Uncaught Exception vulnerability)
 - Refactored investigation class to be qualitynotification to reflect a base class for future extensions
 - Renamed notification table to investigation_notification to be able to understand the difference of notification source
+
 ### Removed
 - Not needed enum params in UpdateInvestigationRequest